requirements.txt
twscrape==0.1.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-api-python-client==2.24.0
googleapis-common-protos==1.53.0
httplib2==0.20.1
oauthlib==3.1.1
pytz==2021.3
python-dotenv==0.19.1
